pyproject.toml
[project]
name = "matrix"
version = "0.1.0"
description = "Personal agent harness"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.7",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "pyyaml>=6.0",
]

[project.scripts]
matrix = "matrix.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["matrix"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]
